requirements.txt
mysql-connector-python
pandas
numpy
sklearn
xgboost
scipy